chore: move cli_config_wizard_test.py to scripts/test_flow, rename shared.py -> test_helpers.py
Source: owner request
Type: refactor
